API, A new feature of checking peer's liveness
- A new API has been added that allows clients to request a liveness
check for a session using IkeSession's requestLinvenessCheck API.
Liveness check can be done using on-demand DPD (dead peer detection)
or with a valid incoming packet in busyState.
- A new API onLivenessStatusChanged of IkeSessionCallback, which
allows clients to know the liveness status, has also been added. When
the liveness check begins, 'START' if liveness check is started or
'ONGOING' if liveness check is already in progress are notified. and
'SUCCESS' or 'FAILURE' are notified as a result. Note that 'FAILURE'
status is followed by session closure.
- The getLivenessRetransmissionTimeoutsMillis API has been added to
IkeSessionParam and the setLivenessRetransmissionTimeoutsMillis API
also has been added to IkeSessionParam.Builder to perform more
aggressive detection for on-demand DPD.

Bug: 299343224
Bug: 290539988
Bug: 301180369
Test: atest FrameworksIkeTests CtsIkeTestCases
Test: On device functional test.
Change-Id: I84c24832a731e27239c254a1e4e77dcef728d2f5