Fix IPv6 IKE Socket Selection when NAT Detection is Disabled
This fix addresses an issue where IKE over IPv6 would incorrectly switch
to a UDP encapsulation socket when NAT detection is disabled. This
misalignment caused issues with IPsec, as UDP encapsulation of ESP is
not supported for IPv6. The fix ensures that the correct non-UDP
encapsulation socket is maintained, preventing any compatibility
problems.

Bug: 362788262
Test: atest FrameworksIkeTests
Flag: EXEMPT bugfix
Change-Id: I5634a15d2b891624acbf94f1f30101789025154f